Refactor into Server Directory
I like to organize my code better than the default directory structure.
This is a persnickety habit, to be sure. That said, I like to be able to
find my code in a structure that feels logical to me.

- create `server` directory
- create `server/models` and `server/controllers`
- move `routes` and `views` to `server`
- modify code to reference new file locations
